refator: renamed folders and files
Here are the changes that I made:

1. Removed all the redundant names in modules and submodules. The namespace is supposed to tell you what the module is for not the module itself. Instead of `doe_project.doe_geometry` it is now `doe.geometry`.

2. Move inp related files to the main module. I don't think we need a submodule for that.

3. Moved all the reference files and other assets under a single folder called assets.
