+++ assets/reference_inp/boundary_conditions.md
