+++ assets/reference_inp/minimum_input.md
