+++ assets/reference_inp/simple_example_dir_ref.jpg
