+++ dragonfly_doe2/doe/assemblies/idf_2_inp_opaque.py
