+++ dragonfly_doe2/inp.py
@@ -1,7 +1,7 @@
-from doe_project.doe_model import INPRoom, INPStory, INPModel
+from doe.model import INPRoom, INPStory, INPModel
 
 
-# TODO: refactor like hb ies from the classes in doe_project to do a one-shot write out from the classes
+# TODO: refactor like hb ies from the classes in doe to do a one-shot write out from the classes
 """
 Notes:
 Make agnostic 'lookup classes' to use for ease of itteration
